requirements.txt
Flask==2.0.2
Flask-SQLAlchemy
numpy
openai
pandas
python-docx
python-dotenv==0.19.1
requests
sklearn
SQLAlchemy==1.4
torch
transformers
